refactor(actix): main, lib 분리
Closes #13
